complete signup

+++ src/screens/SignupScreen.js
@@ -7,24 +7,31 @@ import { Link, useLocation, useNavigate } from 'react-router-dom';
 import { Store } from '../Store';
 import getError from '../utils';
 
-const SigninScreen = () => {
+const SignupScreen = () => {
     const navigate = useNavigate();
     const { search } = useLocation();
     const redirectInUrl = new URLSearchParams(search).get('redirect');
     const redirect = redirectInUrl ? redirectInUrl : '/';
 
+    const [name, setName] = useState('');
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
+    const [confirmPassword, setConfirmPassword] = useState('');
 
     const { state, dispatch: ctxDispatch } = useContext(Store);
     const { userInfo } = state;
 
     const submitHandler = async (e) => {
         e.preventDefault();
+        if (password !== confirmPassword) {
+            toast.error('Your passwords do not match. Try again..');
+            return;
+        }
         try {
-            const { data } = await Axios.post('/api/users/signin', {
+            const { data } = await Axios.post('/api/users/signup', {
+                name,
                 email,
-                password,
+                password
             });
             ctxDispatch({ type: 'USER_SIGNIN', payload: data })
             localStorage.setItem('userInfo', JSON.stringify(data));
@@ -46,28 +53,39 @@ const SigninScreen = () => {
     return (
         <Container className='small-container'>
             <Helmet>
-                <title>Sign In</title>
+                <title>Sign Up</title>
             </Helmet>
-            <h1 className="my-3">Sign In</h1>
+            <h1 className="my-3">Sign Up</h1>
             <Form onSubmit={submitHandler}>
+                <Form.Group className='mb-3' controlId='name'>
+                    <Form.Label>Name</Form.Label>
+                    <Form.Control required onChange={(e) => setName(e.target.value)} />
+                </Form.Group>
+
                 <Form.Group className='mb-3' controlId='email'>
                     <Form.Label>Email</Form.Label>
                     <Form.Control type='email' required onChange={(e) => setEmail(e.target.value)} />
                 </Form.Group>
+
                 <Form.Group className='mb-3' controlId='password'>
                     <Form.Label>Password</Form.Label>
                     <Form.Control type='password' required onChange={(e) => setPassword(e.target.value)} />
                 </Form.Group>
+
+                <Form.Group className='mb-3' controlId='confirmPassword'>
+                    <Form.Label>Confirm Password</Form.Label>
+                    <Form.Control type='Password' required onChange={(e) => setConfirmPassword(e.target.value)} />
+                </Form.Group>
                 <div className='mb-3'>
-                    <Button type='submit'>Sign In</Button>
+                    <Button type='submit'>Sign Up</Button>
                 </div>
                 <div className='mb-3'>
-                    Don't have an account? {' '}
-                    <Link to={`/signup?redirect=${redirect}`}>Get Started</Link>
+                    Already have an account? {' '}
+                    <Link to={`/signin?redirect=${redirect}`}>Sign In</Link>
                 </div>
             </Form>
         </Container>
     )
 }
 
-export default SigninScreen
+export default SignupScreen
